(OM-43717) Fixed post.install script

+++ source/scripts/omni.isaac.sim.post.install.sh
@@ -4,9 +4,20 @@ SCRIPT_DIR=$(dirname ${BASH_SOURCE})
 
 # Add symlink to Isaac Examples
 pushd ${SCRIPT_DIR}
-ln -s exts/omni.isaac.examples/omni/isaac/examples extension_examples
+if [ ! -L extension_examples ] && [ ! -e extension_examples ]; then
+    ln -s exts/omni.isaac.examples/omni/isaac/examples extension_examples
+    echo "Symlink extension_examples created"
+else
+    echo "Symlink or folder extension_examples exists"
+fi
 popd
+
 # Warm up shader cache
+echo "Warming up cache..."
 ${SCRIPT_DIR}/omni.isaac.sim.warmup.sh
+
 # Install default Python packages 
+echo "Installing Python packages..."
 ${SCRIPT_DIR}/python.sh -m pip install -r ${SCRIPT_DIR}/requirements.txt
+
+echo "Isaac Sim post installation script completed!"
